+++ Task_mech.h
@@ -16,13 +16,13 @@ void task_init(void);
     Configure tasks
     if state > 0 then task with number "nr" is enabled
 */
-int task_config(unsigned int nr, unsigned int state);
+int32 task_config(uint8 nr, uint8 state);
 
 /*
     Instead of using "task_config()" one can also use
     these functions:
 */
-int task_enable(unsigned int nr);
-int task_disable(unsigned int nr);
+int32 task_enable(uint8 nr);
+int32 task_disable(uint8 nr);
 
 #endif  // __TASK_MECH_H
